+++ python/cuml/__init__.py
@@ -44,7 +44,7 @@ from cuml.manifold.umap import UMAP
 
 from cuml.random_projection.random_projection import GaussianRandomProjection, SparseRandomProjection, johnson_lindenstrauss_min_dim
 
-from cuml.preprocessing.utils import train_test_split
+from cuml.preprocessing.model_selection import train_test_split
 
 from ._version import get_versions
 __version__ = get_versions()['version']
